refactor(cluster): remove deferred queue dead code (#1018)
Remove all deferred queue infrastructure from sim/cluster/ — the gateway
queue (#882) superseded this feature, making #899 moot.

- Remove deferredQueue field, isBusy(), promoteDeferred(), DeferredQueueLen()
- Delete cluster_deferred_test.go (preserve TestBatchRequestsNotSerialized
  regression guard for #965 in cluster_test.go)
- Delete specs/004-deferred-queue-batch/ (superseded spec)
- Update CLAUDE.md and invariants.md to remove stale references
- Fix bookkeeping-bug comment to remove deferred-queue framing
