+++ CLAUDE.md
@@ -298,14 +298,12 @@ Request processing pipeline: Arrival → Admission → Routing → WaitQueue →
 ## Recent Changes
 - GAIE-legacy saturation-based admission (#1014): `gaie-legacy` admission policy replicates production llm-d/GAIE admission behavior. Saturation formula: `avg(max(qd/qdThreshold, kvUtil/kvThreshold))` across instances. Non-sheddable requests (priority >= 0) always pass; sheddable requests (priority < 0) rejected when saturation >= 1.0. Defaults: `gaie_qd_threshold=5`, `gaie_kv_threshold=0.8`. Empty pool → saturation=1.0 (conservative). Configured via policy bundle YAML only. Per-tier shed counter (`shedByTier`) now tracks all tier-aware admission rejections unconditionally.
 - Configurable SLO tier priorities (#1013): `SLOPriorityMap` type replaces hardcoded `SLOTierPriority()`. GAIE-compatible defaults: critical=4, standard=3, batch=-1, sheddable=-2, background=-3. `IsSheddable(class)` returns `priority < 0` (matches llm-d `sheddable.go`). Configurable via policy bundle YAML `slo_priorities` in `AdmissionConfig` (e.g., `admission: { slo_priorities: { batch: 0 } }` to make batch non-sheddable). Tenant budget enforcement uses `IsSheddable()` instead of hardcoded priority threshold. `TierShedMinPriority` validation removed (GAIE priorities are unbounded integers).
-- Remove deferred-queue pre-admission intercept (#1012/#1016): Batch/background requests no longer intercepted before admission — they flow through normal admission pipeline. `DeferredQueue` retained for `NodeReadyEvent` lifecycle only. Deferred-queue depth removed from INV-1 conservation equation.
 - TraceV2 ITL timestamps (#992): `blis observe --record-itl` captures per-chunk timestamps for ITL (inter-token latency) calibration. `blis calibrate --itl-data` computes ITL metric (mean chunk-to-chunk delta) alongside TTFT and E2E. ITL data stored in separate CSV (`request_id,chunk_index,timestamp_us`). Backward compatible: ITL is opt-in, TraceV2 format unchanged.
-- fix(workload): inference_perf SLOClass regression (#965): Changed `SLOClass` from `"batch"` to `"standard"` in `ExpandInferencePerfSpec`. Commit `8bc7a48c` introduced a deferred queue that serialized all `batch`-class requests; inference_perf workloads (used by all training experiments) had `SLOClass: "batch"` as a semantically-inert legacy label, inflating TTFT 6–100× after the deferred queue was added. `model_configs/*/config.json` is now checked in for testing and documentation. Golden dataset `testdata/trained_physics_iter29.json` added with 15 iter29 experiments and `TestTrainedPhysics_GoldenDataset` in `sim/cluster/`.
+- fix(workload): inference_perf SLOClass regression (#965): Changed `SLOClass` from `"batch"` to `"standard"` in `ExpandInferencePerfSpec`. `model_configs/*/config.json` is now checked in for testing and documentation. Golden dataset `testdata/trained_physics_iter29.json` added with 15 iter29 experiments and `TestTrainedPhysics_GoldenDataset` in `sim/cluster/`.
 - Cache signal propagation delay (#919): `--cache-signal-delay` flag adds configurable staleness to `precise-prefix-cache` and `no-hit-lru` scorers. When > 0, scorers query periodically-refreshed stale snapshots of each instance's `HashToBlock` map via `StaleCacheIndex`, modeling asynchronous KV event propagation from production llm-d. Default 2s (2,000,000 µs), matching llm-d's `defaultSpeculativeTTL` — the blind spot between routing decision and KV event arrival via ZMQ. Set to 0 for oracle mode (live cache state). INV-7 table updated with cacheQueryFn signal freshness tier.
 - Precise prefix cache scoring (#883): `precise-prefix-cache` and `no-hit-lru` scorers query actual instance KV cache state via `CacheQueryFn` threading through `NewRoutingPolicy`. `GetCachedBlockCount` accessor on `InstanceSimulator`. Cluster layer builds `cacheQueryFn` from instances, including deferred NodePool instances.
 - Gateway queue with saturation-gated dispatch (#882): `SaturationDetector` interface (NeverSaturated, UtilizationDetector, ConcurrencyDetector), `GatewayQueue` with FIFO/Priority dispatch, completion-triggered dispatch, per-request `GatewayQueueDelay` metric, INV-1 conservation extended with `gateway_queue_depth` + `gateway_queue_shed`
 - fix(cluster): pool-authoritative hardware calibration for all backends (#888/#893): `NodePool.gpu_type` overrides `--gpu` for both GPU label and roofline `HWConfig` (TFlopsPeak/BwPeakTBs). `DeploymentConfig.HWConfigByGPU` map supplies per-pool `HardwareCalib`; lookup applied in both sync and deferred (`NodeReadyEvent`) construction paths. `CachedSnapshotProvider.AddInstance` added for dynamic instance registration.
 - Phase 1B-2b: Per-tenant Jain fairness index in simulation output (#812, PR #881): `ComputePerTenantMetrics` + `printPerTenantMetrics` wired into `blis run` and `blis replay`; section absent for untenanted/legacy workloads
-- Phase 1B-2a: Deferred queue for batch/background requests (#810)
 - Phase 1B-1b: Per-tenant fair-share tracking and admission enforcement (#811)
 - Phase 1A: Added node/GPU placement, instance lifecycle, multi-model routing, per-model metrics
