--- specs/004-deferred-queue-batch/checklists/requirements.md
@@ -1,34 +1,0 @@
-# Specification Quality Checklist: Deferred Queue for Batch and Background Requests
-
-**Purpose**: Validate specification completeness and quality before proceeding to planning
-**Created**: 2026-03-26
-**Feature**: [spec.md](../spec.md)
-
-## Content Quality
-
-- [X] No implementation details (languages, frameworks, APIs)
-- [X] Focused on user value and business needs
-- [X] Written for non-technical stakeholders
-- [X] All mandatory sections completed
-
-## Requirement Completeness
-
-- [X] No [NEEDS CLARIFICATION] markers remain
-- [X] Requirements are testable and unambiguous
-- [X] Success criteria are measurable
-- [X] Success criteria are technology-agnostic (no implementation details)
-- [X] All acceptance scenarios are defined
-- [X] Edge cases are identified
-- [X] Scope is clearly bounded
-- [X] Dependencies and assumptions identified
-
-## Feature Readiness
-
-- [X] All functional requirements have clear acceptance criteria
-- [X] User scenarios cover primary flows
-- [X] Feature meets measurable outcomes defined in Success Criteria
-- [X] No implementation details leak into specification
-
-## Notes
-
-All items pass. Spec is ready for `/speckit.plan`.
